Revert "Use the new mangling for reflection."
This needs some changes in lldb.
Disabled for now until lldb supports the new mangling.

This reverts commit 21ba2929432d2cbf4c01e45a9fe83cf059020ca0.
