+++ lib/Basic/Demangle.cpp
@@ -16,8 +16,10 @@
 
 #include "swift/Basic/Demangle.h"
 #include "swift/Basic/Fallthrough.h"
+#ifndef NO_NEW_DEMANGLING
 #include "swift/Basic/Demangler.h"
 #include "swift/Basic/ManglingMacros.h"
+#endif
 #include "swift/Strings.h"
 #include "swift/Basic/LLVM.h"
 #include "swift/Basic/Punycode.h"
@@ -323,10 +325,12 @@ public:
   ///
   /// \return true if the mangling succeeded
   NodePointer demangleTopLevel() {
+#ifndef NO_NEW_DEMANGLING
     if (Mangled.str().startswith(MANGLING_PREFIX_STR)) {
       NewMangling::Demangler D(Mangled.str());
       return D.demangleTopLevel();
     }
+#endif
     if (!Mangled.nextIf("_T"))
       return nullptr;
 
@@ -2340,9 +2344,11 @@ swift::Demangle::isSwiftSymbol(const char *mangledName,
   if (Name.startswith("_T"))
     return true;
 
+#ifndef NO_NEW_DEMANGLING
   // The new mangling.
   if (Name.startswith(MANGLING_PREFIX_STR))
     return true;
+#endif // !NO_NEW_DEMANGLING
   return false;
 }
 
@@ -2350,6 +2356,7 @@ bool
 swift::Demangle::isThunkSymbol(const char *mangledName,
                                size_t mangledNameLength) {
   StringRef Name(mangledName, mangledNameLength);
+#ifndef NO_NEW_DEMANGLING
   if (Name.startswith(MANGLING_PREFIX_STR)) {
     // First do a quick check
     if (Name.endswith("TA") ||  // partial application forwarder
@@ -2375,6 +2382,7 @@ swift::Demangle::isThunkSymbol(const char *mangledName,
     }
     return false;
   }
+#endif // !NO_NEW_DEMANGLING
 
   if (Name.startswith("_T")) {
     // Old mangling.
@@ -2400,8 +2408,8 @@ NodePointer
 swift::Demangle::demangleTypeAsNode(const char *MangledName,
                                     size_t MangledNameLength,
                                     const DemangleOptions &Options) {
-  NewMangling::Demangler D(StringRef(MangledName, MangledNameLength));
-  return D.demangleType();
+  Demangler demangler(StringRef(MangledName, MangledNameLength));
+  return demangler.demangleTypeName();
 }
 
 namespace {
