+++ lib/Basic/Remangler.cpp
@@ -1687,126 +1687,11 @@ void Remangler::mangleSILBoxImmutableField(Node *node) {
 } // anonymous namespace
 
 /// The top-level interface to the remangler.
-std::string Demangle::mangleNode(const NodePointer &node) {
+std::string Demangle::mangleNodeNew(const NodePointer &node) {
   if (!node) return "";
 
   DemanglerPrinter printer;
   Remangler(printer).mangle(node.get());
 
   return std::move(printer).str();
-}
-
-static bool isInSwiftModule(Node *node) {
-  auto context = node->begin()->get();
-  return (context->getKind() == Node::Kind::Module &&
-          context->getText() == STDLIB_NAME);
-};
-
-bool Demangle::mangleStandardSubstitution(Node *node, DemanglerPrinter &Out) {
-  // Look for known substitutions.
-  switch (node->getKind()) {
-#define SUCCESS_IF_IS(VALUE, EXPECTED, SUBSTITUTION)            \
-do {                                                        \
-if ((VALUE) == (EXPECTED)) {                              \
-Out << SUBSTITUTION;                                    \
-return true;                                            \
-}                                                         \
-} while (0)
-#define SUCCESS_IF_TEXT_IS(EXPECTED, SUBSTITUTION)              \
-SUCCESS_IF_IS(node->getText(), EXPECTED, SUBSTITUTION)
-#define SUCCESS_IF_DECLNAME_IS(EXPECTED, SUBSTITUTION)          \
-SUCCESS_IF_IS(node->getChild(1)->getText(), EXPECTED, SUBSTITUTION)
-
-    case Node::Kind::Module:
-      SUCCESS_IF_TEXT_IS(STDLIB_NAME, "s");
-      SUCCESS_IF_TEXT_IS(MANGLING_MODULE_OBJC, "So");
-      SUCCESS_IF_TEXT_IS(MANGLING_MODULE_C, "SC");
-      break;
-    case Node::Kind::Structure:
-      if (isInSwiftModule(node)) {
-        SUCCESS_IF_DECLNAME_IS("Array", "Sa");
-        SUCCESS_IF_DECLNAME_IS("Bool", "Sb");
-        SUCCESS_IF_DECLNAME_IS("UnicodeScalar", "Sc");
-        SUCCESS_IF_DECLNAME_IS("Double", "Sd");
-        SUCCESS_IF_DECLNAME_IS("Float", "Sf");
-        SUCCESS_IF_DECLNAME_IS("Int", "Si");
-        SUCCESS_IF_DECLNAME_IS("UnsafeRawPointer", "SV");
-        SUCCESS_IF_DECLNAME_IS("UnsafeMutableRawPointer", "Sv");
-        SUCCESS_IF_DECLNAME_IS("UnsafePointer", "SP");
-        SUCCESS_IF_DECLNAME_IS("UnsafeMutablePointer", "Sp");
-        SUCCESS_IF_DECLNAME_IS("UnsafeBufferPointer", "SR");
-        SUCCESS_IF_DECLNAME_IS("UnsafeMutableBufferPointer", "Sr");
-        SUCCESS_IF_DECLNAME_IS("String", "SS");
-        SUCCESS_IF_DECLNAME_IS("UInt", "Su");
-      }
-      break;
-    case Node::Kind::Enum:
-      if (isInSwiftModule(node)) {
-        SUCCESS_IF_DECLNAME_IS("Optional", "Sq");
-        SUCCESS_IF_DECLNAME_IS("ImplicitlyUnwrappedOptional", "SQ");
-      }
-      break;
-
-    default:
-      break;
-
-#undef SUCCESS_IF_DECLNAME_IS
-#undef SUCCESS_IF_TEXT_IS
-#undef SUCCESS_IF_IS
-  }
-  return false;
-}
-
-bool Demangle::isSpecialized(Node *node) {
-  switch (node->getKind()) {
-    case Node::Kind::BoundGenericStructure:
-    case Node::Kind::BoundGenericEnum:
-    case Node::Kind::BoundGenericClass:
-      return true;
-
-    case Node::Kind::Structure:
-    case Node::Kind::Enum:
-    case Node::Kind::Class: {
-      Node *parentOrModule = node->getChild(0).get();
-      if (isSpecialized(parentOrModule))
-        return true;
-
-      return false;
-    }
-
-    default:
-      return false;
-  }
-}
-
-NodePointer Demangle::getUnspecialized(Node *node) {
-  switch (node->getKind()) {
-    case Node::Kind::Structure:
-    case Node::Kind::Enum:
-    case Node::Kind::Class: {
-      NodePointer result = NodeFactory::create(node->getKind());
-      NodePointer parentOrModule = node->getChild(0);
-      if (isSpecialized(parentOrModule.get()))
-        result->addChild(getUnspecialized(parentOrModule.get()));
-      else
-        result->addChild(parentOrModule);
-      result->addChild(node->getChild(1));
-      return result;
-    }
-
-    case Node::Kind::BoundGenericStructure:
-    case Node::Kind::BoundGenericEnum:
-    case Node::Kind::BoundGenericClass: {
-      NodePointer unboundType = node->getChild(0);
-      assert(unboundType->getKind() == Node::Kind::Type);
-      NodePointer nominalType = unboundType->getChild(0);
-      if (isSpecialized(nominalType.get()))
-        return getUnspecialized(nominalType.get());
-      else
-        return nominalType;
-    }
-      
-    default:
-      unreachable("bad nominal type kind");
-  }
 }
