+++ lib/IRGen/GenReflection.cpp
@@ -31,7 +31,6 @@
 #include "IRGenModule.h"
 #include "Linking.h"
 #include "LoadableTypeInfo.h"
-#include "IRGenMangler.h"
 
 using namespace swift;
 using namespace irgen;
@@ -214,19 +213,22 @@ protected:
     // for the purposes of reflection metadata
     assert(!type->hasArchetype() && "Forgot to map typeref out of context");
 
+    Mangle::Mangler mangler(/*DWARFMangling*/false,
+                            /*usePunyCode*/ true,
+                            /*OptimizeProtocolNames*/ false);
+    mangler.setModuleContext(ModuleContext);
+    
     // TODO: As a compatibility hack, mangle single-field boxes with the legacy
     // mangling in reflection metadata.
-    bool isSingleFieldOfBox = false;
     auto boxTy = dyn_cast<SILBoxType>(type);
     if (boxTy && boxTy->getLayout()->getFields().size() == 1) {
       GenericContextScope scope(IGM, Context);
-      type = boxTy->getFieldLoweredType(IGM.getSILModule(), 0);
-      isSingleFieldOfBox = true;
+      mangler.mangleLegacyBoxType(
+        boxTy->getFieldLoweredType(IGM.getSILModule(), 0));
+    } else {
+      mangler.mangleType(type, 0);
     }
-    IRGenMangler mangler;
-    std::string MangledStr = mangler.mangleTypeForReflection(type,
-                                            ModuleContext, isSingleFieldOfBox);
-    auto mangledName = IGM.getAddrOfStringForTypeRef(MangledStr);
+    auto mangledName = IGM.getAddrOfStringForTypeRef(mangler.finalize());
     addRelativeAddress(mangledName);
   }
