+++ stdlib/public/Reflection/TypeRefBuilder.cpp
@@ -76,7 +76,7 @@ lookupSuperclass(const std::string &MangledTypeName) {
   // Superclasses are recorded as a special associated type named 'super'
   // on the 'AnyObject' protocol.
   return lookupTypeWitness(MangledTypeName, "super",
-                           ProtocolTypeRef::create(*this, "s9AnyObject_p"));
+                           ProtocolTypeRef::create(*this, "Ps9AnyObject_"));
 }
 
 const TypeRef * TypeRefBuilder::
