+++ stdlib/public/runtime/Demangle.cpp
@@ -10,8 +10,9 @@
 //
 //===----------------------------------------------------------------------===//
 
+#define NO_NEW_DEMANGLING
+
 #include "../../../lib/Basic/Demangle.cpp"
-#include "../../../lib/Basic/Demangler.cpp"
 #include "../../../lib/Basic/ManglingUtils.cpp"
 #include "../../../lib/Basic/Punycode.cpp"
 #include "swift/Runtime/Metadata.h"
