+++ stdlib/public/runtime/Remangle.cpp
@@ -10,5 +10,5 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "../../../lib/Basic/Remangler.cpp"
+#include "../../../lib/Basic/Remangle.cpp"
 #include "../../../lib/Basic/PunycodeUTF8.cpp"
